requirements.txt
streamlit
python-dotenv
google-generativeai
google-api-core
requests
beautifulsoup4
duckduckgo-search
